pyproject.toml
[project]
name = "biosyn-kb"
version = "0.1.0"
description = "Web-search-driven biosynthesis knowledge base agent"
readme = "README.md"
requires-python = ">=3.9"
authors = [{name = "Biosyn KB"}]
dependencies = [
  "httpx>=0.27.0",
  "pydantic>=2.6.0",
  "pydantic-settings>=2.2.0",
  "beautifulsoup4>=4.12.0",
  "tldextract>=5.1.2",
  "PyYAML>=6.0.1",
  "trafilatura>=1.7.0",
  "SQLAlchemy>=2.0.30",
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "Jinja2>=3.1.2",
  "chromadb>=0.5.23"
  ,"python-multipart>=0.0.9"
  ,"reportlab>=4.2.2"
  ,"sentence-transformers>=2.2.2"
  ,"playwright>=1.47.0"
  ,"pdfminer.six>=20221105"
]

[project.scripts]
biosyn-kb = "biosyn_kb.cli:main"
biosyn-kb-web = "biosyn_kb.webapp.main:run"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
